backend/requirements.txt
flask
flask-cors
PyJWT
pytest
